+++ lib/ace/tcp/endpoint.ex
@@ -23,7 +23,7 @@ defmodule Ace.TCP.Endpoint do
     {:mode, :binary},
 
     # Received packets are delineated on each new line.
-    {:packet, :line},
+    {:packet, :raw},
 
     # Set the socket to execute in passive mode.
     # The process must explicity receive incoming data by calling `TCP.recv/2`